+++ app/[locale]/calibrate/page.tsx
@@ -6,6 +6,7 @@ import {
   useCallback,
   useEffect,
   useReducer,
+  useRef,
   useState,
 } from "react";
 import { FullScreen, useFullScreenHandle } from "react-full-screen";
@@ -14,7 +15,7 @@ import CalibrationCanvas from "@/_components/calibration-canvas";
 import Draggable from "@/_components/draggable";
 import Header from "@/_components/header";
 import {
-  getCenterPoint,
+  getCalibrationCenterPoint,
   getPerspectiveTransformFromPoints,
 } from "@/_lib/geometry";
 import isValidPDF from "@/_lib/is-valid-pdf";
@@ -44,6 +45,7 @@ import CalibrationContext, {
   getCalibrationContext,
   getIsInvalidatedCalibrationContext,
   getIsInvalidatedCalibrationContextWithPointerEvent,
+  logCalibrationContextDifferences,
 } from "@/_lib/calibration-context";
 import WarningIcon from "@/_icons/warning-icon";
 import PdfViewer from "@/_components/pdf-viewer";
@@ -53,10 +55,11 @@ import stitchSettingsReducer from "@/_reducers/stitchSettingsReducer";
 import { StitchSettings } from "@/_lib/interfaces/stitch-settings";
 import Tooltip from "@/_components/tooltip/tooltip";
 import { IconButton } from "@/_components/buttons/icon-button";
-import FullscreenExitIcon from "@/_icons/fullscreen-exit-icon";
+import FullScreenExitIcon from "@/_icons/full-screen-exit-icon";
 import { Layers } from "@/_lib/layers";
 import useLayers from "@/_hooks/use-layers";
 import ExpandMoreIcon from "@/_icons/expand-more-icon";
+import FullScreenIcon from "@/_icons/full-screen-icon";
 
 const defaultStitchSettings = {
   columnCount: 1,
@@ -91,6 +94,7 @@ export default function Page() {
   const [calibrationTransform, setCalibrationTransform] = useState<Matrix>(
     Matrix.identity(3, 3),
   );
+  const [restoreTransform, setRestoreTransform] = useState<Matrix | null>(null);
   const [pageCount, setPageCount] = useState<number>(0);
   const [unitOfMeasure, setUnitOfMeasure] = useState(IN);
   const { layers, dispatchLayersAction } = useLayers(file?.name ?? "default");
@@ -102,6 +106,10 @@ export default function Page() {
   const [layoutHeight, setLayoutHeight] = useState<number>(0);
   const [lineThickness, setLineThickness] = useState<number>(0);
   const [measuring, setMeasuring] = useState<boolean>(false);
+  const [magnifying, setMagnifying] = useState<boolean>(false);
+  const [zoomedOut, setZoomedOut] = useState<boolean>(false);
+  const [menusHidden, setMenusHidden] = useState<boolean>(false);
+  const [isIdle, setIsIdle] = useState(false);
 
   const [menuStates, setMenuStates] = useState<MenuStates>(
     getDefaultMenuStates(),
@@ -112,7 +120,21 @@ export default function Page() {
   const [fullScreenTooltipVisible, setFullScreenTooltipVisible] =
     useState(true);
 
+  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
+
   const t = useTranslations("Header");
+
+  const IDLE_TIMEOUT = 8000;
+
+  function resetIdle() {
+    setIsIdle(false);
+    if (timeoutRef.current) {
+      clearTimeout(timeoutRef.current);
+    }
+    timeoutRef.current = setTimeout(() => {
+      setIsIdle(true);
+    }, IDLE_TIMEOUT);
+  }
 
   function getDefaultPoints() {
     const { innerWidth, innerHeight } = window;
@@ -203,6 +225,10 @@ export default function Page() {
     if (!file) {
       return;
     }
+    setZoomedOut(false);
+    setRestoreTransform(null);
+    setMagnifying(false);
+    setMeasuring(false);
     setPageCount(0); // Reset page count while loading
     const key = `stitchSettings:${file.name ?? "default"}`;
     const stitchSettingsString = localStorage.getItem(key);
@@ -292,6 +318,7 @@ export default function Page() {
   }, []);
 
   function handlePointerDown(e: React.PointerEvent) {
+    resetIdle();
     if (fullScreenTooltipVisible) {
       setFullScreenTooltipVisible(false);
     }
@@ -309,10 +336,24 @@ export default function Page() {
             fullScreenHandle.active,
           )
         ) {
+          logCalibrationContextDifferences(expected, fullScreenHandle.active);
           setCalibrationValidated(false);
         }
       }
     }
+
+    setMenusHidden(false);
+  }
+
+  function handlePointerUp() {
+    if (magnifying) {
+      setMagnifying(false);
+    }
+  }
+
+  function handlePointerMove() {
+    resetIdle();
+    setMenusHidden(false);
   }
 
   useEffect(() => {
@@ -320,6 +361,10 @@ export default function Page() {
       !isCalibrating && !calibrationValidated;
     setShowCalibrationAlert(projectingWithInvalidContext);
   }, [isCalibrating, calibrationValidated]);
+
+  useEffect(() => {
+    setMenusHidden(isIdle && !isCalibrating);
+  }, [isIdle, isCalibrating]);
 
   useEffect(() => {
     const interval = setInterval(() => {
@@ -347,8 +392,10 @@ export default function Page() {
   return (
     <main
       onPointerDown={handlePointerDown}
+      onPointerUp={handlePointerUp}
+      onPointerMove={handlePointerMove}
       ref={noZoomRefCallback}
-      className={`${isDarkTheme(displaySettings.theme) && "dark bg-black"} w-screen h-screen absolute overflow-hidden touch-none`}
+      className={`${menusHidden && "cursor-none"} ${isDarkTheme(displaySettings.theme) && "dark bg-black"} w-screen h-screen absolute overflow-hidden touch-none`}
     >
       <div className="bg-white dark:bg-black dark:text-white w-screen h-screen ">
         <FullScreen
@@ -376,9 +423,9 @@ export default function Page() {
                   }
                 >
                   {fullScreenHandle.active ? (
-                    <FullscreenExitIcon ariaLabel={t("fullscreen")} />
+                    <FullScreenIcon ariaLabel={t("fullscreen")} />
                   ) : (
-                    <FullscreenExitIcon ariaLabel={t("fullscreenExit")} />
+                    <FullScreenExitIcon ariaLabel={t("fullscreenExit")} />
                   )}
                 </IconButton>
               </Tooltip>
@@ -417,14 +464,30 @@ export default function Page() {
               measuring={measuring}
               setMeasuring={setMeasuring}
               file={file}
-              gridCenter={getCenterPoint(+width, +height, unitOfMeasure)}
+              gridCenter={getCalibrationCenterPoint(
+                +width,
+                +height,
+                unitOfMeasure,
+              )}
             >
               <Draggable
-                className={`absolute`}
+                className={`absolute ${menusHidden && "!cursor-none"} `}
                 perspective={perspective}
                 isCalibrating={isCalibrating}
                 unitOfMeasure={unitOfMeasure}
                 calibrationTransform={calibrationTransform}
+                magnifying={magnifying}
+                setRestoreTransform={setRestoreTransform}
+                restoreTransform={restoreTransform}
+                zoomedOut={zoomedOut}
+                setZoomedOut={setZoomedOut}
+                layoutWidth={layoutWidth}
+                layoutHeight={layoutHeight}
+                calibrationCenter={getCalibrationCenterPoint(
+                  +width,
+                  +height,
+                  unitOfMeasure,
+                )}
               >
                 <PdfViewer
                   file={file}
@@ -451,7 +514,7 @@ export default function Page() {
             </MeasureCanvas>
 
             <menu
-              className={`absolute ${menuStates.nav ? "top-0" : "-top-16"} w-screen`}
+              className={`absolute w-screen pointer-events-auto ${visible(!menusHidden)} ${menuStates.nav ? "top-0" : "-top-16"} pointer-events-none`}
             >
               <Header
                 isCalibrating={isCalibrating}
@@ -483,7 +546,6 @@ export default function Page() {
                     updateLocalSettings(newSettings);
                   }
                 }}
-                pageCount={pageCount}
                 layoutWidth={layoutWidth}
                 layoutHeight={layoutHeight}
                 lineThickness={lineThickness}
@@ -499,8 +561,14 @@ export default function Page() {
                 }}
                 setCalibrationValidated={setCalibrationValidated}
                 fullScreenTooltipVisible={fullScreenTooltipVisible}
+                magnifying={magnifying}
+                setMagnifying={setMagnifying}
+                zoomedOut={zoomedOut}
+                setZoomedOut={setZoomedOut}
               />
-              <menu className={`${visible(!isCalibrating && file !== null)}`}>
+              <menu
+                className={`${visible(!isCalibrating && file !== null)} p-0`}
+              >
                 <StitchMenu
                   className={`${menuStates.stitch && menuStates.nav ? "opacity-100 block" : "opacity-0 hidden"}`}
                   setShowMenu={(showMenu) =>
@@ -511,6 +579,8 @@ export default function Page() {
                   pageCount={pageCount}
                   file={file}
                   layers={layers}
+                  menuStates={menuStates}
+                  setMenuStates={setMenuStates}
                 />
                 <LayerMenu
                   visible={menuStates.layers}
@@ -523,7 +593,7 @@ export default function Page() {
               </menu>
             </menu>
             <IconButton
-              className={`!p-1 m-0 border-2 border-black dark:border-white absolute ${menuStates.nav ? "-top-16" : "top-2"} left-1/4 focus:ring-0`}
+              className={`${visible(!menusHidden)} !p-1 m-0 border-2 border-black dark:border-white absolute ${menuStates.nav ? "-top-16" : "top-2"} left-1/4 focus:ring-0`}
               onClick={() => setMenuStates({ ...menuStates, nav: true })}
             >
               <ExpandMoreIcon ariaLabel={t("menuShow")} />