+++ app/_components/draggable.tsx
@@ -1,5 +1,11 @@
 import { Matrix } from "ml-matrix";
-import { ReactNode, useState, useRef, useEffect } from "react";
+import {
+  ReactNode,
+  useState,
+  useEffect,
+  SetStateAction,
+  Dispatch,
+} from "react";
 
 import { toMatrix3d, transformPoint, translate } from "@/_lib/geometry";
 import { Point } from "@/_lib/point";
@@ -19,6 +25,14 @@ export default function Draggable({
   unitOfMeasure,
   calibrationTransform,
   className,
+  magnifying,
+  restoreTransform,
+  setRestoreTransform,
+  zoomedOut,
+  setZoomedOut,
+  layoutWidth,
+  layoutHeight,
+  calibrationCenter,
 }: {
   children: ReactNode;
   perspective: Matrix;
@@ -26,13 +40,19 @@ export default function Draggable({
   unitOfMeasure: string;
   calibrationTransform: Matrix;
   className: string;
+  magnifying: boolean;
+  setRestoreTransform: Dispatch<SetStateAction<Matrix | null>>;
+  restoreTransform: Matrix | null;
+  zoomedOut: boolean;
+  setZoomedOut: Dispatch<SetStateAction<boolean>>;
+  layoutWidth: number;
+  layoutHeight: number;
+  calibrationCenter: Point;
 }) {
   const [dragStart, setDragStart] = useState<Point | null>(null);
   const [transformStart, setTransformStart] = useState<Matrix | null>(null);
-  const [isIdle, setIsIdle] = useState(false);
-  const [matrix3d, setMatrix3d] = useState<string>("");
 
-  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
+  const [matrix3d, setMatrix3d] = useState<string>("");
 
   const transform = useTransformContext();
   const transformer = useTransformerContext();
@@ -49,18 +69,6 @@ export default function Draggable({
     setMatrix3d(toMatrix3d(calibrationTransform.mmul(transform)));
   }, [transform, calibrationTransform]);
 
-  const IDLE_TIMEOUT = 1500;
-
-  function resetIdle() {
-    setIsIdle(false);
-    if (timeoutRef.current) {
-      clearTimeout(timeoutRef.current);
-    }
-    timeoutRef.current = setTimeout(() => {
-      setIsIdle(true);
-    }, IDLE_TIMEOUT);
-  }
-
   function handleOnEnd(): void {
     setDragStart(null);
     setTransformStart(null);
@@ -70,7 +78,6 @@ export default function Draggable({
     const p = { x: e.clientX, y: e.clientY };
 
     if (e.pointerType === "mouse") {
-      resetIdle();
       /* If we aren't currently dragging, ignore the mouse move event */
       if (dragStart === null) {
         return;
@@ -94,26 +101,59 @@ export default function Draggable({
   function handleOnStart(e: React.PointerEvent): void {
     const p = { x: e.clientX, y: e.clientY };
     const pt = transformPoint(p, perspective);
-    setDragStart(pt);
-    setTransformStart(transform.clone());
+    if (magnifying) {
+      setRestoreTransform(transform.clone());
+      transformer.magnify(5, pt);
+    } else if (restoreTransform !== null) {
+      setRestoreTransform(null);
+      setZoomedOut(false);
+      transformer.zoomIn(p, calibrationCenter);
+    } else {
+      setDragStart(pt);
+      setTransformStart(transform.clone());
+    }
   }
 
-  let cursorMode = `${dragStart !== null ? "grabbing" : "grab"}`;
-  let viewportCursorMode = `${dragStart !== null ? "grabbing" : "default"}`;
+  const cursorMode = `${dragStart !== null ? "grabbing" : magnifying || zoomedOut ? "zoom-in" : "grab"}`;
+  const viewportCursorMode = `${dragStart !== null ? "grabbing" : "default"}`;
 
-  /* If we aren't dragging and the idle timer has set isIdle
-   * to true, hide the cursor */
-  if (dragStart === null && isIdle) {
-    cursorMode = "none";
-    viewportCursorMode = "none";
-  }
+  useEffect(() => {
+    if (zoomedOut && restoreTransform === null) {
+      setRestoreTransform(transform.clone());
+      transformer.zoomOut(layoutWidth, layoutHeight, calibrationTransform);
+    }
+  }, [
+    zoomedOut,
+    setZoomedOut,
+    transformer,
+    layoutWidth,
+    layoutHeight,
+    calibrationTransform,
+    transform,
+    setRestoreTransform,
+    restoreTransform,
+  ]);
+
+  useEffect(() => {
+    if (!magnifying && !zoomedOut) {
+      if (restoreTransform !== null) {
+        transformer.setLocalTransform(restoreTransform);
+        setRestoreTransform(null);
+      }
+    }
+  }, [
+    magnifying,
+    zoomedOut,
+    restoreTransform,
+    setRestoreTransform,
+    transformer,
+  ]);
 
   return (
     <div
       tabIndex={0}
       className={`${className ?? ""} select-none absolute top-0 ${visible(!isCalibrating)} bg-white dark:bg-black transition-all duration-500 w-screen h-screen`}
       onPointerMove={handleMove}
-      onMouseEnter={resetIdle}
       onMouseUp={handleOnEnd}
       style={{
         cursor: viewportCursorMode,
@@ -131,7 +171,7 @@ export default function Draggable({
         <div
           className={"absolute"}
           style={{
-            transform: `${matrix3d}`,
+            transform: `${zoomedOut ? `scale(${transform.get(0, 0)})` : matrix3d}`,
             transformOrigin: "0 0",
           }}
         >