+++ app/_components/header.tsx
@@ -26,15 +26,13 @@ import {
 } from "@/_lib/display-settings";
 import { CM, IN } from "@/_lib/unit";
 import RecenterIcon from "@/_icons/recenter-icon";
-import { getCenterPoint } from "@/_lib/geometry";
+import { getCalibrationCenterPoint } from "@/_lib/geometry";
 import { visible } from "@/_components/theme/css-functions";
 import { IconButton } from "@/_components/buttons/icon-button";
 import { DropdownCheckboxIconButton } from "@/_components/buttons/dropdown-checkbox-icon-button";
 import Tooltip from "@/_components/tooltip/tooltip";
-import FullscreenExitIcon from "@/_icons/fullscreen-exit-icon";
 import ExpandLessIcon from "@/_icons/expand-less-icon";
 import LineWeightIcon from "@/_icons/line-weight-icon";
-import FlexWrapIcon from "@/_icons/flex-wrap-icon";
 import { useKeyDown } from "@/_hooks/use-key-down";
 import { KeyCode } from "@/_lib/key-code";
 import { MenuStates } from "@/_lib/menu-states";
@@ -52,6 +50,10 @@ import { useTransformerContext } from "@/_hooks/use-transform-context";
 import { DropdownIconButton } from "./buttons/dropdown-icon-button";
 import MarkAndMeasureIcon from "@/_icons/mark-and-measure-icon";
 import FlippedPatternIcon from "@/_icons/flipped-pattern-icon";
+import MagnifyIcon from "@/_icons/magnify-icon";
+import ZoomOutIcon from "@/_icons/zoom-out-icon";
+import FullSceenExitIcon from "@/_icons/full-screen-exit-icon";
+import FullScreenIcon from "@/_icons/full-screen-icon";
 
 export default function Header({
   isCalibrating,
@@ -67,7 +69,6 @@ export default function Header({
   setUnitOfMeasure,
   displaySettings,
   setDisplaySettings,
-  pageCount,
   layoutWidth,
   layoutHeight,
   lineThickness,
@@ -80,6 +81,10 @@ export default function Header({
   setShowingMovePad,
   setCalibrationValidated,
   fullScreenTooltipVisible,
+  magnifying,
+  setMagnifying,
+  zoomedOut,
+  setZoomedOut,
 }: {
   isCalibrating: boolean;
   setIsCalibrating: Dispatch<SetStateAction<boolean>>;
@@ -94,7 +99,6 @@ export default function Header({
   setUnitOfMeasure: (newUnit: string) => void;
   displaySettings: DisplaySettings;
   setDisplaySettings: (newDisplaySettings: DisplaySettings) => void;
-  pageCount: number;
   layoutWidth: number;
   layoutHeight: number;
   lineThickness: number;
@@ -107,6 +111,10 @@ export default function Header({
   setShowingMovePad: Dispatch<SetStateAction<boolean>>;
   setCalibrationValidated: Dispatch<SetStateAction<boolean>>;
   fullScreenTooltipVisible: boolean;
+  magnifying: boolean;
+  setMagnifying: Dispatch<SetStateAction<boolean>>;
+  zoomedOut: boolean;
+  setZoomedOut: Dispatch<SetStateAction<boolean>>;
 }) {
   const [calibrationAlert, setCalibrationAlert] = useState("");
   const transformer = useTransformerContext();
@@ -152,20 +160,27 @@ export default function Header({
   }
 
   const handleRotate90 = () => {
-    transformer.rotate(getCenterPoint(+width, +height, unitOfMeasure), 90);
+    transformer.rotate(
+      getCalibrationCenterPoint(+width, +height, unitOfMeasure),
+      90,
+    );
   };
 
   const handleFlipHorizontal = () => {
-    transformer.flipHorizontal(getCenterPoint(+width, +height, unitOfMeasure));
+    transformer.flipHorizontal(
+      getCalibrationCenterPoint(+width, +height, unitOfMeasure),
+    );
   };
 
   const handleFlipVertical = () => {
-    transformer.flipVertical(getCenterPoint(+width, +height, unitOfMeasure));
+    transformer.flipVertical(
+      getCalibrationCenterPoint(+width, +height, unitOfMeasure),
+    );
   };
 
   const handleRecenter = () => {
     transformer.recenter(
-      getCenterPoint(+width, +height, unitOfMeasure),
+      getCalibrationCenterPoint(+width, +height, unitOfMeasure),
       layoutWidth,
       layoutHeight,
     );
@@ -243,7 +258,15 @@ export default function Header({
 
   useKeyDown(() => {
     setMeasuring(!measuring);
+  }, [KeyCode.KeyL]);
+
+  useKeyDown(() => {
+    setMagnifying(!magnifying);
   }, [KeyCode.KeyM]);
+
+  useKeyDown(() => {
+    setZoomedOut(!zoomedOut);
+  }, [KeyCode.KeyZ]);
 
   return (
     <>
@@ -270,13 +293,13 @@ export default function Header({
         </ModalActions>
       </Modal>
       <header
-        className={`relative z-10 bg-white dark:bg-black left-0 w-full border-b dark:border-gray-700 transition-all duration-500 h-16 flex items-center ${menuStates.nav ? "translate-y-0" : "-translate-y-16"}`}
+        className={`relative z-10 bg-opacity-80 dark:bg-opacity-70 bg-white dark:bg-black left-0 w-full border-b dark:border-gray-700 transition-all duration-500 h-16 flex items-center ${menuStates.nav ? "translate-y-0" : "-translate-y-16"}`}
       >
         <nav
           className="mx-auto flex max-w-7xl items-center justify-between p-2 lg:px-8 w-full"
           aria-label="Global"
         >
-          <div className="flex items-center gap-2">
+          <div className="flex items-center gap-1">
             <Tooltip
               description={
                 fullScreenHandle.active ? t("fullscreenExit") : t("fullscreen")
@@ -291,43 +314,21 @@ export default function Header({
                 }
               >
                 {fullScreenHandle.active ? (
-                  <FullscreenExitIcon ariaLabel={t("fullscreen")} />
+                  <FullScreenIcon ariaLabel={t("fullscreen")} />
                 ) : (
-                  <FullscreenExitIcon ariaLabel={t("fullscreenExit")} />
+                  <FullSceenExitIcon ariaLabel={t("fullscreenExit")} />
                 )}
               </IconButton>
             </Tooltip>
-            <Tooltip description={t("menuHide")}>
+            <Tooltip
+              description={t("menuHide")}
+              className={visible(isCalibrating)}
+            >
               <IconButton
                 className={`!p-1 border-2 border-black dark:border-white`}
                 onClick={() => setMenuStates({ ...menuStates, nav: false })}
               >
                 <ExpandLessIcon ariaLabel={t("menuHide")} />
-              </IconButton>
-            </Tooltip>
-            <Tooltip
-              description={
-                menuStates.stitch
-                  ? t("stitchMenuHide")
-                  : pageCount === 0
-                    ? t("stitchMenuDisabled")
-                    : t("stitchMenuShow")
-              }
-            >
-              <IconButton
-                disabled={pageCount === 0}
-                onClick={() =>
-                  setMenuStates({ ...menuStates, stitch: !menuStates.stitch })
-                }
-                className={`${menuStates.stitch ? "!bg-gray-300 dark:!bg-gray-600" : ""} ${visible(!isCalibrating)}`}
-              >
-                <FlexWrapIcon
-                  ariaLabel={
-                    menuStates.stitch
-                      ? t("stitchMenuHide")
-                      : t("stitchMenuShow")
-                  }
-                />
               </IconButton>
             </Tooltip>
             <Tooltip description={t("invertColor")}>
@@ -381,7 +382,7 @@ export default function Header({
               />
             )}
           </div>
-          <div className={`flex items-center gap-2 ${visible(isCalibrating)}`}>
+          <div className={`flex items-center gap-1 ${visible(isCalibrating)}`}>
             <div className="flex gap-1">
               <InlineInput
                 className="relative flex flex-col"
@@ -443,28 +444,29 @@ export default function Header({
               </IconButton>
             </Tooltip>
           </div>
-          <div className={`flex items-center gap-2 ${visible(!isCalibrating)}`}>
+          <div className={`flex items-center gap-1 ${visible(!isCalibrating)}`}>
             <Tooltip description={t("flipHorizontal")}>
-              <IconButton onClick={handleFlipHorizontal}>
+              <IconButton onClick={handleFlipHorizontal} disabled={zoomedOut}>
                 <FlipVerticalIcon ariaLabel={t("flipHorizontal")} />
               </IconButton>
             </Tooltip>
             <Tooltip description={t("flipVertical")}>
-              <IconButton onClick={handleFlipVertical}>
+              <IconButton onClick={handleFlipVertical} disabled={zoomedOut}>
                 <FlipHorizontalIcon ariaLabel={t("flipVertical")} />
               </IconButton>
             </Tooltip>
             <Tooltip description={t("rotate90")}>
-              <IconButton onClick={handleRotate90}>
+              <IconButton onClick={handleRotate90} disabled={zoomedOut}>
                 <Rotate90DegreesCWIcon ariaLabel={t("rotate90")} />
               </IconButton>
             </Tooltip>
             <Tooltip description={t("recenter")}>
               <IconButton
+                disabled={zoomedOut}
                 onClick={() => {
                   transformer.reset();
                   transformer.recenter(
-                    getCenterPoint(+width, +height, unitOfMeasure),
+                    getCalibrationCenterPoint(+width, +height, unitOfMeasure),
                     layoutWidth,
                     layoutHeight,
                   );
@@ -481,12 +483,25 @@ export default function Header({
                 <MarkAndMeasureIcon ariaLabel={t("measure")} />
               </IconButton>
             </Tooltip>
+            <Tooltip description={t("magnify")}>
+              <IconButton
+                onClick={() => setMagnifying(!magnifying)}
+                disabled={zoomedOut}
+              >
+                <MagnifyIcon ariaLabel={t("magnify")} />
+              </IconButton>
+            </Tooltip>
+            <Tooltip description={t("zoomOut")}>
+              <IconButton onClick={() => setZoomedOut(!zoomedOut)}>
+                <ZoomOutIcon ariaLabel={t("zoomOut")} />
+              </IconButton>
+            </Tooltip>
           </div>
-          <div className="flex items-center gap-4">
+          <div className="flex items-center gap-2">
             <label
               className={`${visible(
                 !isCalibrating,
-              )} flex gap-2 items-center outline outline-purple-600 text-purple-600 focus:ring-2 focus:outline-none focus:ring-blue-300 dark:focus:ring-blue-800  hover:bg-purple-600 hover:text-white font-medium rounded-lg text-sm px-2 py-1.5 hover:bg-none text-center`}
+              )} flex gap-2 items-center outline outline-purple-600 text-purple-600 focus:ring-2 focus:outline-none focus:ring-blue-300 dark:focus:ring-blue-800  hover:bg-purple-600 dark:bg-black bg-white hover:text-white font-medium rounded-lg text-sm px-2 py-1.5 hover:bg-none text-center`}
             >
               <FileInput
                 accept="application/pdf"
@@ -503,7 +518,7 @@ export default function Header({
             >
               {isCalibrating ? t("project") : t("calibrate")}
             </button>
-            <Tooltip description={t("info")}>
+            <Tooltip description={t("info")} className={visible(isCalibrating)}>
               <IconButton href="/">
                 <InfoIcon ariaLabel={t("info")} />
               </IconButton>